requirements.txt
Flask
gunicorn
sqlalchemy
pymysql
mysql-connector-python
python-decouple